Move maintenance window resource into domain package
