+++ internal/provider/maintenancewindow/resource_test.go
@@ -1,4 +1,4 @@
-package provider
+package maintenancewindow
 
 import (
 	"context"
